requirements.txt
langchain_openai
langchain
pandas
numpy 
streamlit
python-dotenv
langchain_community
langchain_core
